Completed homework 2
Had to look at the solution to see how making the typedValidator worked. It appears the params need to be put into the typedValidator function and the dataum type is inserted in the wrap = Scripts.wrapValidator @() @!--Insert type of dataum here--!

Also, I have a question about homework1: It appears the times do not work as expected.  It appears beneficiary1 cannot grab until strictly AFTER the deadline is passed and beneficary2 can grab at the end of the deadline, but I don't understand how this is the case based on how the code is written.  I would like this functionality explained...